Fixed: wrong local variable type in weather system.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@3430 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Weather.cpp
@@ -120,7 +120,7 @@ bool Weather::ReGenerate()
 
     // remember old values
     uint32 old_type = m_type;
-    uint32 old_grade = m_grade;
+    float old_grade = m_grade;
 
     m_grade = rand_norm();
     time_t gtime = sWorld.GetGameTime();
